+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### Unreleased
+- Document the support for templated links according to the HAL JSON specification
+
 ### v 3.0.0 (27 Feb 2020)
 - When a relation name is present in both embedded and links, embedded will take precedence, because it may contain more information.
 - Remove embeddedListName option again because it is currently unused, incomplete and creates many corner cases